Finishing rewrite
