+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "timeline-map",
-  "version": "0.0.7",
+  "version": "0.0.9",
   "description": "TimeLine Map Component from MapsForAll",
   "main": "dist/TimeLineMap.js",
   "scripts": {
