+++ src/style.css
@@ -1,2 +1,5 @@
-.map-control {  }
-.date-control { margin: 30px; }
+.map-control {
+}
+.date-control {
+  margin: 30px 30px 50px 30px;
+}
