+++ webpack.config.js
@@ -1,27 +1,27 @@
 /* global __dirname, require, module*/
 const config = {
-  output: {
-    filename: 'TimeLineMap.js',
-    libraryTarget: 'umd',
-    umdNamedDefine: true
-  },
-  module: {
-    rules: [
+    output: {
+        filename: 'TimeLineMap.js',
+        libraryTarget: 'umd',
+        umdNamedDefine: true
+    },
+    module: {
+        rules: [
       { test: /\.js$/, loader: 'babel-loader' },
-      {
-        test: /\.css$/,
-        use: [
+            {
+                test: /\.css$/,
+                use: [
           { loader: 'style-loader' },
-          {
-            loader: 'css-loader',
-            options: {
-              modules: false
+                    {
+                        loader: 'css-loader',
+                        options: {
+                            modules: false
+                        }
+                    }
+                ]
             }
-          }
         ]
-      }
-    ]
-  }
+    }
 };
 
 module.exports = config;
